Move interface comms to new class
