+++ nightingale_dispatcher/src/nightingale_dispatcher/handoff_task.py
@@ -2,10 +2,10 @@
 
 import rospy
 
-from nightingale_dispatcher.task import Task
+from nightingale_dispatcher.task_with_interface_comms import TaskWithInterfaceComms
 
 
-class HandoffTask(Task):
+class HandoffTask(TaskWithInterfaceComms):
     # Once arrived at the bedside, pass of items to patient
     # TODO define parameters
     def __init__(self, *args, **kwargs):
